Remove lib/cloud dependencies from lib/services (#41740)
Importing cloud specific dependencies causes bloat in client tools
and will be prohibited with a linter in the future. This moves all
remaining imports into lib/srv/discovery in a similar manner that
was done for lib/cloud/gcp in https://github.com/gravitational/teleport/pull/41678.
