+++ lib/services/server.go
@@ -21,19 +21,14 @@ package services
 import (
 	"encoding/json"
 	"fmt"
-	"net"
 	"time"
 
-	"github.com/aws/aws-sdk-go-v2/aws"
-	ec2types "github.com/aws/aws-sdk-go-v2/service/ec2/types"
-	ec2v1 "github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/google/go-cmp/cmp"
 	"github.com/gravitational/trace"
 
 	apidefaults "github.com/gravitational/teleport/api/defaults"
 	"github.com/gravitational/teleport/api/types"
 	apiutils "github.com/gravitational/teleport/api/utils"
-	libaws "github.com/gravitational/teleport/lib/cloud/aws"
 	"github.com/gravitational/teleport/lib/defaults"
 	"github.com/gravitational/teleport/lib/utils"
 )
@@ -45,9 +40,6 @@ const (
 	OnlyTimestampsDifferent = iota
 	// Different means that some fields are different
 	Different = iota
-
-	// defaultSSHPort is the default port for the OpenSSH Service.
-	defaultSSHPort = "22"
 )
 
 // CompareServers compares two provided servers.
@@ -430,72 +422,4 @@ func MarshalServers(s []types.Server) ([]byte, error) {
 func NodeHasMissedKeepAlives(s types.Server) bool {
 	serverExpiry := s.Expiry()
 	return serverExpiry.Before(time.Now().Add(apidefaults.ServerAnnounceTTL - (apidefaults.ServerKeepAliveTTL() * 2)))
-}
-
-// NewAWSNodeFromEC2Instance creates a Node resource from an EC2 Instance.
-// It has a pre-populated spec which contains info that is not available in the ec2.Instance object.
-func NewAWSNodeFromEC2Instance(instance ec2types.Instance, awsCloudMetadata *types.AWSInfo) (types.Server, error) {
-	labels := libaws.TagsToLabels(instance.Tags)
-	if labels == nil {
-		labels = make(map[string]string)
-	}
-	libaws.AddMetadataLabels(labels, awsCloudMetadata.AccountID, awsCloudMetadata.Region)
-
-	instanceID := aws.ToString(instance.InstanceId)
-	labels[types.AWSInstanceIDLabel] = instanceID
-	labels[types.AWSAccountIDLabel] = awsCloudMetadata.AccountID
-
-	awsCloudMetadata.InstanceID = instanceID
-	awsCloudMetadata.VPCID = aws.ToString(instance.VpcId)
-	awsCloudMetadata.SubnetID = aws.ToString(instance.SubnetId)
-
-	if aws.ToString(instance.PrivateIpAddress) == "" {
-		return nil, trace.BadParameter("private ip address is required from ec2 instance")
-	}
-	// Address requires the Port.
-	// We use the default port for the OpenSSH daemon.
-	addr := net.JoinHostPort(aws.ToString(instance.PrivateIpAddress), defaultSSHPort)
-
-	server, err := types.NewEICENode(
-		types.ServerSpecV2{
-			Hostname: aws.ToString(instance.PrivateDnsName),
-			Addr:     addr,
-			CloudMetadata: &types.CloudMetadata{
-				AWS: awsCloudMetadata,
-			},
-		},
-		labels,
-	)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	return server, nil
-}
-
-// NewAWSNodeFromEC2v1Instance creates a Node resource from an EC2 Instance.
-// It has a pre-populated spec which contains info that is not available in the ec2.Instance object.
-// Uses AWS SDK Go V1
-func NewAWSNodeFromEC2v1Instance(instance ec2v1.Instance, awsCloudMetadata *types.AWSInfo) (types.Server, error) {
-	server, err := NewAWSNodeFromEC2Instance(ec2InstanceV1ToV2(instance), awsCloudMetadata)
-	return server, trace.Wrap(err)
-}
-
-func ec2InstanceV1ToV2(instance ec2v1.Instance) ec2types.Instance {
-	tags := make([]ec2types.Tag, 0, len(instance.Tags))
-	for _, tag := range instance.Tags {
-		tags = append(tags, ec2types.Tag{
-			Key:   tag.Key,
-			Value: tag.Value,
-		})
-	}
-
-	return ec2types.Instance{
-		InstanceId:       instance.InstanceId,
-		VpcId:            instance.VpcId,
-		SubnetId:         instance.SubnetId,
-		PrivateIpAddress: instance.PrivateIpAddress,
-		PrivateDnsName:   instance.PrivateDnsName,
-		Tags:             tags,
-	}
 }
