+++ lib/srv/db/cloud/resource_checker_url_aws.go
@@ -32,7 +32,7 @@ import (
 	apiawsutils "github.com/gravitational/teleport/api/utils/aws"
 	"github.com/gravitational/teleport/lib/cloud"
 	cloudaws "github.com/gravitational/teleport/lib/cloud/aws"
-	"github.com/gravitational/teleport/lib/services"
+	"github.com/gravitational/teleport/lib/srv/discovery/common"
 )
 
 func (c *urlChecker) checkAWS(describeCheck, basicEndpointCheck checkDatabaseFunc) checkDatabaseFunc {
@@ -101,7 +101,7 @@ func (c *urlChecker) checkRDSCluster(ctx context.Context, database types.Databas
 	if err != nil {
 		return trace.Wrap(err)
 	}
-	databases, err := services.NewDatabasesFromRDSCluster(rdsCluster, []*rds.DBInstance{})
+	databases, err := common.NewDatabasesFromRDSCluster(rdsCluster, []*rds.DBInstance{})
 	if err != nil {
 		c.log.Warnf("Could not convert RDS cluster %q to database resources: %v.",
 			aws.StringValue(rdsCluster.DBClusterIdentifier), err)
@@ -211,7 +211,7 @@ func (c *urlChecker) checkElastiCache(ctx context.Context, database types.Databa
 	if err != nil {
 		return trace.Wrap(err)
 	}
-	databases, err := services.NewDatabasesFromElastiCacheReplicationGroup(cluster, nil)
+	databases, err := common.NewDatabasesFromElastiCacheReplicationGroup(cluster, nil)
 	if err != nil {
 		return trace.Wrap(err)
 	}
@@ -254,7 +254,7 @@ func (c *urlChecker) checkOpenSearch(ctx context.Context, database types.Databas
 		return trace.BadParameter("expect 1 domain but got %v", domains.DomainStatusList)
 	}
 
-	databases, err := services.NewDatabasesFromOpenSearchDomain(domains.DomainStatusList[0], nil)
+	databases, err := common.NewDatabasesFromOpenSearchDomain(domains.DomainStatusList[0], nil)
 	if err != nil {
 		return trace.Wrap(err)
 	}
