+++ lib/srv/db/cloud/resource_checker_url_aws_test.go
@@ -35,7 +35,7 @@ import (
 	apiawsutils "github.com/gravitational/teleport/api/utils/aws"
 	"github.com/gravitational/teleport/lib/cloud"
 	"github.com/gravitational/teleport/lib/cloud/mocks"
-	"github.com/gravitational/teleport/lib/services"
+	"github.com/gravitational/teleport/lib/srv/discovery/common"
 )
 
 func TestURLChecker_AWS(t *testing.T) {
@@ -49,64 +49,64 @@ func TestURLChecker_AWS(t *testing.T) {
 
 	// RDS.
 	rdsInstance := mocks.RDSInstance("rds-instance", region, nil)
-	rdsInstanceDB, err := services.NewDatabaseFromRDSInstance(rdsInstance)
+	rdsInstanceDB, err := common.NewDatabaseFromRDSInstance(rdsInstance)
 	require.NoError(t, err)
 	rdsCluster := mocks.RDSCluster("rds-cluster", region, nil,
 		mocks.WithRDSClusterReader,
 		mocks.WithRDSClusterCustomEndpoint("my-custom"),
 	)
-	rdsClusterDBs, err := services.NewDatabasesFromRDSCluster(rdsCluster, []*rds.DBInstance{})
+	rdsClusterDBs, err := common.NewDatabasesFromRDSCluster(rdsCluster, []*rds.DBInstance{})
 	require.NoError(t, err)
 	require.Len(t, rdsClusterDBs, 3) // Primary, reader, custom.
 	testCases = append(testCases, append(rdsClusterDBs, rdsInstanceDB)...)
 
 	// RDS Proxy.
 	rdsProxy := mocks.RDSProxy("rds-proxy", region, "some-vpc")
-	rdsProxyDB, err := services.NewDatabaseFromRDSProxy(rdsProxy, nil)
+	rdsProxyDB, err := common.NewDatabaseFromRDSProxy(rdsProxy, nil)
 	require.NoError(t, err)
 	rdsProxyCustomEndpoint := mocks.RDSProxyCustomEndpoint(rdsProxy, "my-custom", region)
-	rdsProxyCustomEndpointDB, err := services.NewDatabaseFromRDSProxyCustomEndpoint(rdsProxy, rdsProxyCustomEndpoint, nil)
+	rdsProxyCustomEndpointDB, err := common.NewDatabaseFromRDSProxyCustomEndpoint(rdsProxy, rdsProxyCustomEndpoint, nil)
 	require.NoError(t, err)
 	testCases = append(testCases, rdsProxyDB, rdsProxyCustomEndpointDB)
 
 	// Redshift.
 	redshiftCluster := mocks.RedshiftCluster("redshift-cluster", region, nil)
-	redshiftClusterDB, err := services.NewDatabaseFromRedshiftCluster(redshiftCluster)
+	redshiftClusterDB, err := common.NewDatabaseFromRedshiftCluster(redshiftCluster)
 	require.NoError(t, err)
 	testCases = append(testCases, redshiftClusterDB)
 
 	// Redshift Serverless.
 	redshiftServerlessWorkgroup := mocks.RedshiftServerlessWorkgroup("redshift-serverless", region)
-	redshiftServerlessDB, err := services.NewDatabaseFromRedshiftServerlessWorkgroup(redshiftServerlessWorkgroup, nil)
+	redshiftServerlessDB, err := common.NewDatabaseFromRedshiftServerlessWorkgroup(redshiftServerlessWorkgroup, nil)
 	require.NoError(t, err)
 	redshiftServerlessVPCEndpoint := mocks.RedshiftServerlessEndpointAccess(redshiftServerlessWorkgroup, "vpc-endpoint", region)
-	redshiftServerlessVPCEndpointDB, err := services.NewDatabaseFromRedshiftServerlessVPCEndpoint(redshiftServerlessVPCEndpoint, redshiftServerlessWorkgroup, nil)
+	redshiftServerlessVPCEndpointDB, err := common.NewDatabaseFromRedshiftServerlessVPCEndpoint(redshiftServerlessVPCEndpoint, redshiftServerlessWorkgroup, nil)
 	require.NoError(t, err)
 	testCases = append(testCases, redshiftServerlessDB, redshiftServerlessVPCEndpointDB)
 
 	// ElastiCache.
 	elastiCacheCluster := mocks.ElastiCacheCluster("elasticache", region, mocks.WithElastiCacheReaderEndpoint)
-	elastiCacheClusterDBs, err := services.NewDatabasesFromElastiCacheNodeGroups(elastiCacheCluster, nil)
+	elastiCacheClusterDBs, err := common.NewDatabasesFromElastiCacheNodeGroups(elastiCacheCluster, nil)
 	require.NoError(t, err)
 	require.Len(t, elastiCacheClusterDBs, 2) // Primary, reader.
 	elastiCacheClusterConfigurationMode := mocks.ElastiCacheCluster("elasticache-configuration", region, mocks.WithElastiCacheConfigurationEndpoint)
-	elastiCacheClusterConfigurationModeDB, err := services.NewDatabaseFromElastiCacheConfigurationEndpoint(elastiCacheClusterConfigurationMode, nil)
+	elastiCacheClusterConfigurationModeDB, err := common.NewDatabaseFromElastiCacheConfigurationEndpoint(elastiCacheClusterConfigurationMode, nil)
 	require.NoError(t, err)
 	testCases = append(testCases, append(elastiCacheClusterDBs, elastiCacheClusterConfigurationModeDB)...)
 
 	// MemoryDB.
 	memoryDBCluster := mocks.MemoryDBCluster("memorydb", region)
-	memoryDBClusterDB, err := services.NewDatabaseFromMemoryDBCluster(memoryDBCluster, nil)
+	memoryDBClusterDB, err := common.NewDatabaseFromMemoryDBCluster(memoryDBCluster, nil)
 	require.NoError(t, err)
 	testCases = append(testCases, memoryDBClusterDB)
 
 	// OpenSearch.
 	openSearchDomain := mocks.OpenSearchDomain("opensearch", region, mocks.WithOpenSearchCustomEndpoint("custom.com"))
-	openSearchDBs, err := services.NewDatabasesFromOpenSearchDomain(openSearchDomain, nil)
+	openSearchDBs, err := common.NewDatabasesFromOpenSearchDomain(openSearchDomain, nil)
 	require.NoError(t, err)
 	require.Len(t, openSearchDBs, 2) // Primary, custom.
 	openSearchVPCDomain := mocks.OpenSearchDomain("opensearch-vpc", region, mocks.WithOpenSearchVPCEndpoint("vpc"))
-	openSearchVPCDomainDBs, err := services.NewDatabasesFromOpenSearchDomain(openSearchVPCDomain, nil)
+	openSearchVPCDomainDBs, err := common.NewDatabasesFromOpenSearchDomain(openSearchVPCDomain, nil)
 	require.NoError(t, err)
 	require.Len(t, openSearchVPCDomainDBs, 1)
 	testCases = append(testCases, append(openSearchDBs, openSearchVPCDomainDBs...)...)
