+++ lib/srv/discovery/common/server_test.go
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package services
+package common
 
 import (
 	"testing"
