+++ lib/srv/discovery/fetchers/db/aws_elasticache.go
@@ -29,7 +29,6 @@ import (
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/cloud"
 	libcloudaws "github.com/gravitational/teleport/lib/cloud/aws"
-	"github.com/gravitational/teleport/lib/services"
 	"github.com/gravitational/teleport/lib/srv/discovery/common"
 )
 
@@ -63,12 +62,12 @@ func (f *elastiCachePlugin) GetDatabases(ctx context.Context, cfg *awsFetcherCon
 
 	var eligibleClusters []*elasticache.ReplicationGroup
 	for _, cluster := range clusters {
-		if !services.IsElastiCacheClusterSupported(cluster) {
+		if !libcloudaws.IsElastiCacheClusterSupported(cluster) {
 			cfg.Log.Debugf("ElastiCache cluster %q is not supported. Skipping.", aws.StringValue(cluster.ReplicationGroupId))
 			continue
 		}
 
-		if !services.IsElastiCacheClusterAvailable(cluster) {
+		if !libcloudaws.IsElastiCacheClusterAvailable(cluster) {
 			cfg.Log.Debugf("The current status of ElastiCache cluster %q is %q. Skipping.",
 				aws.StringValue(cluster.ReplicationGroupId),
 				aws.StringValue(cluster.Status))
@@ -115,9 +114,9 @@ func (f *elastiCachePlugin) GetDatabases(ctx context.Context, cfg *awsFetcherCon
 			}
 		}
 
-		extraLabels := services.ExtraElastiCacheLabels(cluster, tags, allNodes, allSubnetGroups)
+		extraLabels := common.ExtraElastiCacheLabels(cluster, tags, allNodes, allSubnetGroups)
 
-		if dbs, err := services.NewDatabasesFromElastiCacheReplicationGroup(cluster, extraLabels); err != nil {
+		if dbs, err := common.NewDatabasesFromElastiCacheReplicationGroup(cluster, extraLabels); err != nil {
 			cfg.Log.Infof("Could not convert ElastiCache cluster %q to database resources: %v.",
 				aws.StringValue(cluster.ReplicationGroupId), err)
 		} else {
