+++ lib/srv/discovery/fetchers/db/azure_dbserver.go
@@ -24,7 +24,6 @@ import (
 
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/cloud/azure"
-	"github.com/gravitational/teleport/lib/services"
 	"github.com/gravitational/teleport/lib/srv/discovery/common"
 )
 
@@ -73,7 +72,7 @@ func (p *azureDBServerPlugin) NewDatabaseFromServer(server *azure.DBServer, log 
 		return nil
 	}
 
-	database, err := services.NewDatabaseFromAzureServer(server)
+	database, err := common.NewDatabaseFromAzureServer(server)
 	if err != nil {
 		log.Warnf("Could not convert Azure server %q to database resource: %v.",
 			server.Name,
