+++ lib/srv/discovery/fetchers/db/azure_mysql_flex.go
@@ -25,7 +25,6 @@ import (
 
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/cloud/azure"
-	"github.com/gravitational/teleport/lib/services"
 	"github.com/gravitational/teleport/lib/srv/discovery/common"
 )
 
@@ -57,7 +56,7 @@ func (f *azureMySQLFlexServerFetcher) NewDatabaseFromServer(server *armmysqlflex
 		return nil
 	}
 
-	database, err := services.NewDatabaseFromAzureMySQLFlexServer(server)
+	database, err := common.NewDatabaseFromAzureMySQLFlexServer(server)
 	if err != nil {
 		log.Warnf("Could not convert Azure MySQL server %q to database resource: %v.", azure.StringVal(server.Name), err)
 		return nil
