+++ lib/srv/discovery/fetchers/db/azure_postgres_flex_test.go
@@ -30,7 +30,6 @@ import (
 	azureutils "github.com/gravitational/teleport/api/utils/azure"
 	"github.com/gravitational/teleport/lib/cloud"
 	"github.com/gravitational/teleport/lib/cloud/azure"
-	"github.com/gravitational/teleport/lib/services"
 	"github.com/gravitational/teleport/lib/srv/discovery/common"
 )
 
@@ -89,7 +88,7 @@ func makeAzurePostgresFlexServer(t *testing.T, name, subscription, group, region
 	for _, opt := range opts {
 		opt(server)
 	}
-	database, err := services.NewDatabaseFromAzurePostgresFlexServer(server)
+	database, err := common.NewDatabaseFromAzurePostgresFlexServer(server)
 	require.NoError(t, err)
 	common.ApplyAzureDatabaseNameSuffix(database, types.AzureMatcherPostgres)
 	return server, database
