+++ lib/srv/discovery/fetchers/db/azure_redis.go
@@ -25,7 +25,6 @@ import (
 
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/cloud/azure"
-	"github.com/gravitational/teleport/lib/services"
 	"github.com/gravitational/teleport/lib/srv/discovery/common"
 )
 
@@ -59,7 +58,7 @@ func (p *azureRedisPlugin) NewDatabaseFromServer(server *armredis.ResourceInfo, 
 		return nil
 	}
 
-	database, err := services.NewDatabaseFromAzureRedis(server)
+	database, err := common.NewDatabaseFromAzureRedis(server)
 	if err != nil {
 		log.Warnf("Could not convert Azure Redis server %q to database resource: %v.", azure.StringVal(server.Name), err)
 		return nil
