+++ lib/srv/discovery/fetchers/db/azure_redis_enterprise.go
@@ -25,7 +25,6 @@ import (
 
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/cloud/azure"
-	"github.com/gravitational/teleport/lib/services"
 	"github.com/gravitational/teleport/lib/srv/discovery/common"
 )
 
@@ -66,7 +65,7 @@ func (p *azureRedisEnterprisePlugin) NewDatabaseFromServer(server *azure.RedisEn
 		return nil
 	}
 
-	database, err := services.NewDatabaseFromAzureRedisEnterprise(server.Cluster, server.Database)
+	database, err := common.NewDatabaseFromAzureRedisEnterprise(server.Cluster, server.Database)
 	if err != nil {
 		log.Warnf("Could not convert Azure Redis Enterprise %v to database resource: %v.",
 			server,
